+++ src/index.ts
@@ -111,13 +111,6 @@ Hooks.on('ready', async () => {
             blindroll: 'Blind GM Roll',
             selfroll: 'Self Roll',
         };
-        gameInstance.settings.register<string, string, number>('pf2e-kingmaker-tools', 'averagePartyLevel', {
-            name: 'Average Party Level',
-            default: 1,
-            config: true,
-            type: Number,
-            scope: 'world',
-        });
         gameInstance.settings.register('pf2e-kingmaker-tools', 'showManual', {
             name: 'Show Manual',
             scope: 'world',