+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### Added
+
+* You can now play a playlist when clicking the Begin Rest button on the Camping Sheet
+
 ## [0.36.0] - 2024-02-15
 
 ### Added