+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.38.0] - 2024-02-17
+
+### Changed
+
+* Remove average party level setting, instead take the highest level of a character on the party actor
+
 ## [0.37.0] - 2024-02-16
 
 ### Added